pyproject.toml
[project]
name = "agentic-ai-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "chardet>=7.4.3",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.18",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
]
